backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
face_recognition>=1.3.0
numpy>=1.26
Pillow>=10.3
python-telegram-bot>=21.0
slowapi>=0.1.9
python-dotenv>=1.0
aiofiles>=23.0
